In progress

+++ app/Http/Controllers/Purchase/PurchaseRequestController.php
@@ -353,37 +353,40 @@ class PurchaseRequestController extends Controller
     public function assignVendors(Request $request){
         try{
             $data = $request->all();
-            if($data['is_mail'] == 1){
-                foreach($data['vendor_materials'] as $vendorId => $materialRequestComponentIds){
-                    $purchaseRequestComponentIds = PurchaseRequestComponent::whereIn('material_request_component_id',$materialRequestComponentIds)->pluck('id')->toArray();
-                    PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->whereNotIn('purchase_request_component_id',$purchaseRequestComponentIds)->delete();
+            foreach($data['vendor_materials'] as $vendorId => $materialRequestComponentIds){
+                $purchaseRequestComponentIds = PurchaseRequestComponent::whereIn('material_request_component_id',$materialRequestComponentIds)->pluck('id')->toArray();
+                PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->whereNotIn('purchase_request_component_id',$purchaseRequestComponentIds)->delete();
+                if(array_key_exists('checked_vendor_materials',$data)){
                     if(array_key_exists($vendorId,$data['checked_vendor_materials'])){
                         $vendorInfo = Vendor::findOrFail($vendorId)->toArray();
                         $vendorInfo['materials'] = array();
                     }
-                    $purchaseVendorAssignData = array();
-                    $purchaseVendorAssignData['vendor_id'] = $vendorId;
-                    $iterator = 0;
-                    $jIterator = 0;
-                    $mailInfoData = array();
-                    foreach($materialRequestComponentIds as $materialRequestComponentId){
-                        $materialRequestComponent = MaterialRequestComponents::findOrFail($materialRequestComponentId);
-                        $purchaseVendorAssignData['is_email_sent'] = false;
-                        $purchaseVendorAssignData['purchase_request_component_id'] = $materialRequestComponent->purchaseRequestComponent->id;
-                        // create
-                        $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->where('purchase_request_component_id',$purchaseVendorAssignData['purchase_request_component_id'])->first();
-                        if($purchaseComponentVendorRelation == null){
-                            $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::create($purchaseVendorAssignData);
-                        }
-                        $projectSiteInfo = array();
-                        $projectSiteInfo['project_name'] = $materialRequestComponent->materialRequest->projectSite->project->name;
-                        $projectSiteInfo['project_site_name'] = $materialRequestComponent->materialRequest->projectSite->name;
-                        $projectSiteInfo['project_site_address'] = $materialRequestComponent->materialRequest->projectSite->address;
-                        if($materialRequestComponent->materialRequest->projectSite->city_id == null){
-                            $projectSiteInfo['project_site_city'] = '';
-                        }else{
-                            $projectSiteInfo['project_site_city'] = $materialRequestComponent->materialRequest->projectSite->city->name;
-                        }
+                }
+
+                $purchaseVendorAssignData = array();
+                $purchaseVendorAssignData['vendor_id'] = $vendorId;
+                $iterator = 0;
+                $jIterator = 0;
+                $mailInfoData = array();
+                foreach($materialRequestComponentIds as $materialRequestComponentId){
+                    $materialRequestComponent = MaterialRequestComponents::findOrFail($materialRequestComponentId);
+                    $purchaseVendorAssignData['is_email_sent'] = false;
+                    $purchaseVendorAssignData['purchase_request_component_id'] = $materialRequestComponent->purchaseRequestComponent->id;
+                    // create
+                    $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->where('purchase_request_component_id',$purchaseVendorAssignData['purchase_request_component_id'])->first();
+                    if($purchaseComponentVendorRelation == null){
+                        $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::create($purchaseVendorAssignData);
+                    }
+                    $projectSiteInfo = array();
+                    $projectSiteInfo['project_name'] = $materialRequestComponent->materialRequest->projectSite->project->name;
+                    $projectSiteInfo['project_site_name'] = $materialRequestComponent->materialRequest->projectSite->name;
+                    $projectSiteInfo['project_site_address'] = $materialRequestComponent->materialRequest->projectSite->address;
+                    if($materialRequestComponent->materialRequest->projectSite->city_id == null){
+                        $projectSiteInfo['project_site_city'] = '';
+                    }else{
+                        $projectSiteInfo['project_site_city'] = $materialRequestComponent->materialRequest->projectSite->city->name;
+                    }
+                    if(array_key_exists('checked_vendor_materials',$data)){
                         if(array_key_exists($vendorId,$data['checked_vendor_materials'])){
                             $mailInfoData[$jIterator] = [
                                 'user_id' => Auth::user()->id,
@@ -400,13 +403,16 @@ class PurchaseRequestController extends Controller
                             }
                         }
                     }
-                    if(isset($vendorInfo)){
-                        $pdf = App::make('dompdf.wrapper');
-                        $pdf->loadHTML(view('purchase.purchase-request.pdf.vendor-quotation')->with(compact('vendorInfo','projectSiteInfo')));
-                        $pdfDirectoryPath = env('PURCHASE_VENDOR_ASSIGNMENT_PDF_FOLDER');
-                        $pdfFileName = sha1($vendorId).'.pdf';
-                        $pdfUploadPath = public_path().$pdfDirectoryPath.'/'.$pdfFileName;
-                        $pdfContent = $pdf->stream();
+
+                }
+                if(isset($vendorInfo)){
+                    $pdf = App::make('dompdf.wrapper');
+                    $pdf->loadHTML(view('purchase.purchase-request.pdf.vendor-quotation')->with(compact('vendorInfo','projectSiteInfo')));
+                    $pdfDirectoryPath = env('PURCHASE_VENDOR_ASSIGNMENT_PDF_FOLDER');
+                    $pdfFileName = sha1($vendorId).'.pdf';
+                    $pdfUploadPath = public_path().$pdfDirectoryPath.'/'.$pdfFileName;
+                    $pdfContent = $pdf->stream();
+                    if($data['is_mail'] == 1){
                         if(file_exists($pdfUploadPath)){
                             unlink($pdfUploadPath);
                         }
@@ -424,67 +430,12 @@ class PurchaseRequestController extends Controller
                         PurchaseRequestComponentVendorMailInfo::insert($mailInfoData);
                         PurchaseRequestComponentVendorRelation::whereIn('id',array_column($mailInfoData,'purchase_request_component_vendor_relation_id'))->update(['is_email_sent' => true]);
                         unlink($pdfUploadPath);
-                    }
-                }
-            }else{
-                foreach($data['vendor_materials'] as $vendorId => $materialRequestComponentIds){
-                    $purchaseRequestComponentIds = PurchaseRequestComponent::whereIn('material_request_component_id',$materialRequestComponentIds)->pluck('id')->toArray();
-                    PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->whereNotIn('purchase_request_component_id',$purchaseRequestComponentIds)->delete();
-                    if(array_key_exists($vendorId,$data['checked_vendor_materials'])){
-                        $vendorInfo = Vendor::findOrFail($vendorId)->toArray();
-                        $vendorInfo['materials'] = array();
-                    }
-                    $purchaseVendorAssignData = array();
-                    $purchaseVendorAssignData['vendor_id'] = $vendorId;
-                    $iterator = 0;
-                    $jIterator = 0;
-                    $mailInfoData = array();
-                    foreach($materialRequestComponentIds as $materialRequestComponentId){
-                        $materialRequestComponent = MaterialRequestComponents::findOrFail($materialRequestComponentId);
-                        $purchaseVendorAssignData['is_email_sent'] = false;
-                        $purchaseVendorAssignData['purchase_request_component_id'] = $materialRequestComponent->purchaseRequestComponent->id;
-                        // create
-                        $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::where('vendor_id',$vendorId)->where('purchase_request_component_id',$purchaseVendorAssignData['purchase_request_component_id'])->first();
-                        if($purchaseComponentVendorRelation == null){
-                            $purchaseComponentVendorRelation = PurchaseRequestComponentVendorRelation::create($purchaseVendorAssignData);
-                        }
-                        $projectSiteInfo = array();
-                        $projectSiteInfo['project_name'] = $materialRequestComponent->materialRequest->projectSite->project->name;
-                        $projectSiteInfo['project_site_name'] = $materialRequestComponent->materialRequest->projectSite->name;
-                        $projectSiteInfo['project_site_address'] = $materialRequestComponent->materialRequest->projectSite->address;
-                        if($materialRequestComponent->materialRequest->projectSite->city_id == null){
-                            $projectSiteInfo['project_site_city'] = '';
-                        }else{
-                            $projectSiteInfo['project_site_city'] = $materialRequestComponent->materialRequest->projectSite->city->name;
-                        }
-                        if(array_key_exists($vendorId,$data['checked_vendor_materials'])){
-                            $mailInfoData[$jIterator] = [
-                                'user_id' => Auth::user()->id,
-                                'created_at' => Carbon::now(),
-                                'updated_at' => Carbon::now()
-                            ];
-                            if(in_array($materialRequestComponentId,$data['checked_vendor_materials'][$vendorId])){
-                                $mailInfoData[$jIterator]['purchase_request_component_vendor_relation_id'] = $purchaseComponentVendorRelation->id;
-                                $vendorInfo['materials'][$iterator]['item_name'] = $materialRequestComponent->name;
-                                $vendorInfo['materials'][$iterator]['quantity'] = $materialRequestComponent->quantity;
-                                $vendorInfo['materials'][$iterator]['unit'] = $materialRequestComponent->unit->name;
-                                $iterator++;
-                                $jIterator++;
-                            }
-                        }
-                    }
-                    if(isset($vendorInfo)){
-                        $pdf = App::make('dompdf.wrapper');
-                        $pdf->loadHTML(view('purchase.purchase-request.pdf.vendor-quotation')->with(compact('vendorInfo','projectSiteInfo')));
-                        $pdfDirectoryPath = env('PURCHASE_VENDOR_ASSIGNMENT_PDF_FOLDER');
-                        $pdfFileName = sha1($vendorId).'.pdf';
-                        $pdfUploadPath = public_path().$pdfDirectoryPath.'/'.$pdfFileName;
-                        $pdfContent = $pdf->stream();
+                    }else{
                         return $pdf->stream();
                     }
+
                 }
             }
-
             $request->session()->flash('success','Vendors assigned successfully');
             return redirect('/purchase/purchase-request/edit/p-r-admin-approved/'.$data['purchase_request_id']);
         }catch (\Exception $e){
